fix broken layering
