+++ src/vs/workbench/services/search/node/worker/searchWorkerIpc.ts
@@ -6,7 +6,7 @@
 'use strict';
 
 import { TPromise } from 'vs/base/common/winjs.base';
-import { IChannel } from 'vs/base/parts/ipc/common/ipc';
+import { IChannel } from 'vs/base/parts/ipc/node/ipc';
 import { ISerializedFileMatch } from '../search';
 import { IPatternInfo } from 'vs/platform/search/common/search';
 import { SearchWorker } from './searchWorker';
